requirements.txt
transformers==4.18.0
torch
librosa==0.9.1
t5==0.9.3
note-seq==0.0.3
pretty-midi==0.2.9
einops==0.4.1
ddsp==3.3.4
tensorflow==2.11
tensorflow-text==2.11
protobuf==3.20
numpy==1.22
hydra-core==1.2.0
typing_extensions